Replace receipt_processor.py with refactored implementation
- Delete original receipt_processor.py and rename complete_receipt_processor_v1.py
- Add USD to INR currency conversion with default rate of 83.0
- Fix regex patterns to prevent duplicate amount extraction
- Add required test compatibility methods (_convert_to_inr, _extract_text_tesseract)
- Ensure all 25 test specifications pass
