+++ receipt_processor.py
@@ -1,19 +1,18 @@
 #!/usr/bin/env python3
 """
 Receipt Processing System
-Extracts amounts and categorizes receipts from digital receipts/invoices.
-Supports currency conversion from USD to INR.
+Extracts amount and categorizes receipts from Indian digital receipts/invoices
 """
 
-import argparse
-import csv
 import os
+import csv
 import re
-import traceback
 from pathlib import Path
-from typing import Optional, Tuple, List, Dict, Any, Union
+import argparse
+import traceback
+from typing import Optional, Tuple, List, Dict, Any
 
-# Core dependencies
+# OCR Libraries
 try:
     import pytesseract
     from PIL import Image
@@ -22,14 +21,16 @@ except ImportError:
     HAS_TESSERACT = False
     print("Warning: pytesseract not installed. Install with: pip install pytesseract Pillow")
 
-# Optional dependencies
+# OpenCV for image preprocessing
 try:
     import cv2
     import numpy as np
     HAS_CV2 = True
 except ImportError:
     HAS_CV2 = False
+    print("Info: OpenCV not installed. Install with: pip install opencv-python (for better OCR)")
 
+# PDF processing
 try:
     import fitz  # PyMuPDF
     HAS_PYMUPDF = True
@@ -37,25 +38,21 @@ except ImportError:
     HAS_PYMUPDF = False
     print("Warning: PyMuPDF not installed. Install with: pip install PyMuPDF (for PDF support)")
 
+# Optional: Google Vision API
 try:
     from google.cloud import vision
     HAS_VISION_API = True
 except ImportError:
     HAS_VISION_API = False
 
-try:
-    import requests
-    HAS_REQUESTS = True
-except ImportError:
-    HAS_REQUESTS = False
-
 
 class ReceiptProcessor:
-    """Receipt processor that extracts amounts and categorizes expenses."""
+    """Main class for processing receipt images and extracting expense data."""
     
+    # Supported file formats
     SUPPORTED_FORMATS = {'.jpg', '.jpeg', '.png', '.bmp', '.tiff', '.pdf'}
-    DEFAULT_USD_TO_INR = 83.0
     
+    # Amount extraction patterns
     AMOUNT_PATTERNS = [
         r'₹\s*(\d+(?:,\d{3})*(?:\.\d{1,2})?)',
         r'Rs\.?\s*(\d+(?:,\d{3})*\.\d{1,2})', 
@@ -65,25 +62,34 @@ class ReceiptProcessor:
         r'(?<![a-zA-Z₹$])(\d+(?:,\d{3})+)(?![.\d])'
     ]
     
+    # Keywords to skip when looking for amounts
     SKIP_KEYWORDS = ['credit card', 'card ending', 'order number', 'invoice number', 'phone', 'mobile']
+    
+    # Total indicators
     TOTAL_KEYWORDS = ['grand total', 'total:', 'net total', 'amount due', 'subtotal', 'total', 'amount', 'grand']
+    DEFAULT_USD_TO_INR = 83.0
     
     def __init__(self, use_vision_api: bool = False, debug: bool = False):
-        """Initialize the receipt processor."""
+        """
+        Initialize the Receipt Processor.
+        
+        Args:
+            use_vision_api: Whether to use Google Vision API
+            debug: Enable debug output
+        """
         self.use_vision_api = use_vision_api
         self.debug = debug
         self.categories = self._load_categories()
-        self.usd_to_inr_rate = self._get_exchange_rate()
-        self.vision_client = self._init_vision_client()
-
-    def _init_vision_client(self):
-        """Initialize Google Vision API client if available."""
-        if self.use_vision_api and HAS_VISION_API:
-            return vision.ImageAnnotatorClient()
-        elif self.use_vision_api:
-            print("Warning: Vision API requested but not available")
-        return None
-
+        self.usd_to_inr_rate = self.DEFAULT_USD_TO_INR
+        
+        # Initialize Vision API client if requested and available
+        if use_vision_api and HAS_VISION_API:
+            self.vision_client = vision.ImageAnnotatorClient()
+        else:
+            self.vision_client = None
+            if use_vision_api:
+                print("Warning: Vision API requested but not available")
+    
     def _load_categories(self) -> Dict[str, Dict[str, Any]]:
         """Load predefined expense categories."""
         return {
@@ -129,24 +135,11 @@ class ReceiptProcessor:
                            "microsoft", "google", "aws", "cloud"]
             }
         }
-
-    def _get_exchange_rate(self) -> float:
-        """Get USD to INR exchange rate with API fallback."""
-        if not HAS_REQUESTS:
-            self._debug_print(f"Using default exchange rate: {self.DEFAULT_USD_TO_INR}")
-            return self.DEFAULT_USD_TO_INR
-        
-        try:
-            response = requests.get("https://api.exchangerate-api.com/v4/latest/USD", timeout=5)
-            if response.status_code == 200:
-                rate = response.json()['rates'].get('INR', self.DEFAULT_USD_TO_INR)
-                self._debug_print(f"Got live exchange rate: 1 USD = {rate} INR")
-                return rate
-        except Exception as e:
-            self._debug_print(f"Exchange rate API failed: {e}")
-        
-        self._debug_print(f"Using default exchange rate: {self.DEFAULT_USD_TO_INR}")
-        return self.DEFAULT_USD_TO_INR
+    
+    def _debug_print(self, message: str):
+        """Print debug messages if debug mode is enabled."""
+        if self.debug:
+            print(f"DEBUG: {message}")
 
     def _convert_to_inr(self, amount: float, currency: str) -> float:
         """Convert amount to INR if needed."""
@@ -155,52 +148,131 @@ class ReceiptProcessor:
             self._debug_print(f"Converted ${amount} to ₹{converted:.2f} using rate {self.usd_to_inr_rate}")
             return converted
         return amount
-
-    def _debug_print(self, message: str):
-        """Print debug messages if enabled."""
-        if self.debug:
-            print(f"DEBUG: {message}")
-
-    def extract_text(self, image_path: str) -> str:
-        """Extract text from receipt using available OCR method."""
-        # Try direct PDF text extraction first
-        if str(image_path).lower().endswith('.pdf') and HAS_PYMUPDF:
-            text = self._extract_pdf_text_direct(image_path)
-            if text:
-                return text
-
-        # Try Vision API if available
-        if self.vision_client:
-            text = self._extract_text_vision_api(image_path)
-            if text:
-                return text
-
-        # Fallback to Tesseract
-        if HAS_TESSERACT:
-            return self._extract_text_tesseract(image_path)
+    
+    def pdf_to_images(self, pdf_path: str) -> List[Image.Image]:
+        """Convert PDF to images for OCR."""
+        if not HAS_PYMUPDF:
+            raise ImportError("PyMuPDF not available for PDF processing")
         
-        raise ImportError("No OCR method available")
-
-    def _extract_pdf_text_direct(self, pdf_path: str) -> str:
-        """Extract text directly from PDF."""
         try:
             doc = fitz.open(pdf_path)
-            text_parts = []
+            images = []
+            
             for page_num in range(len(doc)):
                 page = doc.load_page(page_num)
-                text_parts.append(page.get_text())
-            doc.close()
+                # Convert to image with high DPI for better OCR
+                mat = fitz.Matrix(2.0, 2.0)  # 2x zoom for better quality
+                pix = page.get_pixmap(matrix=mat)
+                img_data = pix.tobytes("png")
+                
+                # Convert to PIL Image
+                from io import BytesIO
+                image = Image.open(BytesIO(img_data))
+                images.append(image)
             
-            full_text = '\n'.join(text_parts)
-            if len(full_text.strip()) > 100:
-                self._debug_print(f"Using direct PDF text extraction ({len(full_text)} chars)")
-                return full_text
+            doc.close()
+            return images
+        
         except Exception as e:
-            self._debug_print(f"Direct PDF extraction failed: {e}")
-        return ""
-
-    def _extract_text_vision_api(self, image_path: str) -> str:
+            print(f"Error converting PDF: {e}")
+            return []
+    
+    def preprocess_image(self, image: Image.Image) -> Image.Image:
+        """Preprocess image for better OCR results."""
+        if not HAS_CV2:
+            # Basic preprocessing without OpenCV
+            if image.mode != 'RGB':
+                image = image.convert('RGB')
+            
+            # Resize if too small
+            width, height = image.size
+            if width < 1000:
+                scale_factor = 1500 / width
+                new_size = (int(width * scale_factor), int(height * scale_factor))
+                image = image.resize(new_size, Image.Resampling.LANCZOS)
+            
+            return image
+        
+        # Advanced preprocessing with OpenCV
+        img_array = np.array(image)
+        
+        # Convert to grayscale
+        if len(img_array.shape) == 3:
+            gray = cv2.cvtColor(img_array, cv2.COLOR_RGB2GRAY)
+        else:
+            gray = img_array
+        
+        # Resize if too small
+        height, width = gray.shape
+        if width < 1000:
+            scale_factor = 1500 / width
+            new_width = int(width * scale_factor)
+            new_height = int(height * scale_factor)
+            gray = cv2.resize(gray, (new_width, new_height), interpolation=cv2.INTER_CUBIC)
+        
+        # Apply adaptive thresholding
+        processed = cv2.adaptiveThreshold(gray, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, 
+                                        cv2.THRESH_BINARY, 11, 2)
+        
+        # Denoise
+        processed = cv2.medianBlur(processed, 1)
+        
+        return Image.fromarray(processed)
+    
+    def _extract_text_tesseract(self, image_path: str) -> str:
+        """Private method alias for extract_text_tesseract (for tests)."""
+        return self.extract_text_tesseract(image_path)
+    
+    def extract_text_tesseract(self, image_path: str) -> str:
+        """Extract text using Tesseract OCR."""
+        if not HAS_TESSERACT:
+            raise ImportError("Tesseract not available")
+        
+        try:
+            # Handle PDFs
+            if str(image_path).lower().endswith('.pdf'):
+                if not HAS_PYMUPDF:
+                    return ""
+                
+                images = self.pdf_to_images(image_path)
+                all_text = []
+                
+                for image in images:
+                    image = self.preprocess_image(image)
+                    text = pytesseract.image_to_string(image, lang='eng', config='--psm 6')
+                    all_text.append(text.strip())
+                
+                return '\n'.join(all_text)
+            
+            # Handle regular images
+            image = Image.open(image_path)
+            image = self.preprocess_image(image)
+            
+            # Try multiple PSM modes for better results
+            psm_configs = ['--psm 6', '--psm 3', '--psm 11']
+            texts = []
+            
+            for config in psm_configs:
+                try:
+                    text = pytesseract.image_to_string(image, lang='eng', config=config)
+                    if text.strip():
+                        texts.append(text)
+                        self._debug_print(f"Config '{config}': Found {len(text)} chars")
+                except Exception:
+                    pass
+            
+            # Return the longest extracted text
+            return max(texts, key=len) if texts else ""
+        
+        except Exception as e:
+            self._debug_print(f"Tesseract error: {e}")
+            return ""
+    
+    def extract_text_vision_api(self, image_path: str) -> str:
         """Extract text using Google Vision API."""
+        if not self.vision_client:
+            return ""
+        
         try:
             with open(image_path, 'rb') as image_file:
                 content = image_file.read()
@@ -210,104 +282,45 @@ class ReceiptProcessor:
             texts = response.text_annotations
             
             return texts[0].description.strip() if texts else ""
+        
         except Exception as e:
             self._debug_print(f"Vision API error: {e}")
             return ""
-
-    def _extract_text_tesseract(self, image_path: str) -> str:
-        """Extract text using Tesseract OCR."""
-        try:
-            if str(image_path).lower().endswith('.pdf'):
-                return self._extract_pdf_with_ocr(image_path)
-            
-            # Process regular images
-            image = Image.open(image_path)
-            image = self._preprocess_image(image)
-            
-            # Try multiple PSM configurations
-            configs = ['--psm 6', '--psm 3', '--psm 11']
-            texts = []
-            
-            for config in configs:
-                try:
-                    text = pytesseract.image_to_string(image, lang='eng', config=config)
-                    if text.strip():
-                        texts.append(text)
-                        self._debug_print(f"Config '{config}': Found {len(text)} chars")
-                except Exception:
-                    continue
-            
-            return max(texts, key=len) if texts else ""
-        except Exception as e:
-            self._debug_print(f"Tesseract error: {e}")
-            return ""
-
-    def _extract_pdf_with_ocr(self, pdf_path: str) -> str:
-        """Extract text from PDF using OCR on converted images."""
-        if not HAS_PYMUPDF:
-            return ""
+    
+    def extract_text(self, image_path: str) -> str:
+        """Extract text from receipt image using available OCR method."""
+        # For PDFs, try direct text extraction first
+        if str(image_path).lower().endswith('.pdf'):
+            try:
+                if HAS_PYMUPDF:
+                    doc = fitz.open(image_path)
+                    direct_text = ""
+                    for page_num in range(len(doc)):
+                        page = doc.load_page(page_num)
+                        direct_text += page.get_text() + "\n"
+                    doc.close()
+                    
+                    # If we got substantial text, use it
+                    if len(direct_text.strip()) > 100:
+                        self._debug_print(f"Using direct PDF text extraction ({len(direct_text)} chars)")
+                        return direct_text
+                    else:
+                        self._debug_print("Direct PDF extraction yielded minimal text, falling back to OCR")
+            except Exception as e:
+                self._debug_print(f"Direct PDF extraction failed: {e}, falling back to OCR")
         
-        try:
-            doc = fitz.open(pdf_path)
-            all_text = []
-            
-            for page_num in range(len(doc)):
-                page = doc.load_page(page_num)
-                mat = fitz.Matrix(2.0, 2.0)  # 2x zoom for better quality
-                pix = page.get_pixmap(matrix=mat)
-                
-                from io import BytesIO
-                img_data = BytesIO(pix.tobytes("png"))
-                image = Image.open(img_data)
-                image = self._preprocess_image(image)
-                
-                text = pytesseract.image_to_string(image, lang='eng', config='--psm 6')
-                all_text.append(text.strip())
-            
-            doc.close()
-            return '\n'.join(all_text)
-        except Exception as e:
-            self._debug_print(f"PDF OCR error: {e}")
-            return ""
-
-    def _preprocess_image(self, image: Image.Image) -> Image.Image:
-        """Preprocess image for better OCR results."""
-        if image.mode != 'RGB':
-            image = image.convert('RGB')
+        # Try Vision API first if requested
+        if self.use_vision_api and self.vision_client:
+            text = self.extract_text_vision_api(image_path)
+            if text:
+                return text
         
-        # Resize if too small
-        width, height = image.size
-        if width < 1000:
-            scale_factor = 1500 / width
-            new_size = (int(width * scale_factor), int(height * scale_factor))
-            image = image.resize(new_size, Image.Resampling.LANCZOS)
+        # Fallback to Tesseract
+        if HAS_TESSERACT:
+            return self._extract_text_tesseract(image_path)
         
-        # Apply OpenCV preprocessing if available
-        if HAS_CV2:
-            return self._opencv_preprocess(image)
-        
-        return image
-
-    def _opencv_preprocess(self, image: Image.Image) -> Image.Image:
-        """Apply OpenCV preprocessing for better OCR."""
-        img_array = np.array(image)
-        
-        # Convert to grayscale
-        gray = cv2.cvtColor(img_array, cv2.COLOR_RGB2GRAY) if len(img_array.shape) == 3 else img_array
-        
-        # Resize if needed
-        height, width = gray.shape
-        if width < 1000:
-            scale_factor = 1500 / width
-            new_width, new_height = int(width * scale_factor), int(height * scale_factor)
-            gray = cv2.resize(gray, (new_width, new_height), interpolation=cv2.INTER_CUBIC)
-        
-        # Apply adaptive thresholding and denoising
-        processed = cv2.adaptiveThreshold(gray, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY, 11, 2)
-        processed = cv2.medianBlur(processed, 1)
-        
-        return Image.fromarray(processed)
-
+        raise ImportError("No OCR method available")
+    
     def extract_amount_from_line(self, line: str) -> List[float]:
         """Extract amounts from a single line with currency conversion."""
         amounts = set()
@@ -327,7 +340,7 @@ class ReceiptProcessor:
                     
                     # Detect and convert currency
                     if '$' in line or 'USD' in line.upper():
-                        amount = self._convert_to_inr(amount, 'USD')
+                        amount = amount * self.usd_to_inr_rate
                     
                     if 1 <= amount <= 100000:  # Reasonable range in INR
                         amounts.add(amount)
@@ -336,111 +349,121 @@ class ReceiptProcessor:
                     continue
         
         return sorted(list(amounts))
-
+    
     def extract_amount(self, text: str) -> Optional[float]:
         """Extract the most likely total amount from receipt text."""
+        self._debug_print("Extracting amounts...")
+        
         if not text:
             return None
         
-        self._debug_print("Extracting amounts...")
         amounts_found = []
-        lines = [' '.join(line.split()) for line in text.split('\n')]
+        lines = [' '.join(line.split()) for line in text.split('\n')]  # Normalize whitespace
         
         # Process each line
         for line_num, line in enumerate(lines):
             line = line.strip()
-            if not line or any(keyword in line.lower() for keyword in self.SKIP_KEYWORDS):
+            if not line:
                 continue
             
             line_lower = line.lower()
+            
+            # Skip lines with keywords that typically don't contain the total
+            if any(keyword in line_lower for keyword in self.SKIP_KEYWORDS):
+                continue
+            
+            # Check for money context
             has_currency = any(symbol in line for symbol in ['₹', 'Rs', 'INR', '$'])
             is_total_line = any(keyword in line_lower for keyword in self.TOTAL_KEYWORDS)
             
+            # Extract amounts from this line
             line_amounts = self.extract_amount_from_line(line)
             
             for amount in line_amounts:
-                priority = self._get_amount_priority(line_lower, has_currency, is_total_line)
+                # Assign priority based on context
+                if 'grand total' in line_lower:
+                    priority = 5
+                elif is_total_line:
+                    priority = 4
+                elif has_currency:
+                    priority = 3
+                else:
+                    priority = 1
+                
                 amounts_found.append((amount, priority, line_num, line))
                 self._debug_print(f"Found: ₹{amount} (priority {priority}) in line {line_num}: {line[:50]}...")
         
-        # Check for amounts near total keywords
-        amounts_found.extend(self._find_amounts_near_totals(lines))
+        # Look for amounts near "total" keywords (within 2 lines)
+        for i, line in enumerate(lines):
+            if any(keyword in line.lower() for keyword in ['grand total', 'total:', 'amount due']):
+                # Check nearby lines
+                for offset in [-2, -1, 1, 2]:
+                    if 0 <= i + offset < len(lines):
+                        nearby_line = lines[i + offset]
+                        nearby_amounts = self.extract_amount_from_line(nearby_line)
+                        for amount in nearby_amounts:
+                            # Check if we already have this amount
+                            if not any(a[0] == amount and a[2] == i + offset for a in amounts_found):
+                                amounts_found.append((amount, 4, i + offset, nearby_line))
+                                self._debug_print(f"Found near total: ₹{amount} in line {i + offset}")
         
         if not amounts_found:
             self._debug_print("No amounts found")
             return None
         
-        return self._select_best_amount(amounts_found)
-
-    def _get_amount_priority(self, line_lower: str, has_currency: bool, is_total_line: bool) -> int:
-        """Determine priority for an amount based on context."""
-        if 'grand total' in line_lower:
-            return 5
-        elif is_total_line:
-            return 4
-        elif has_currency:
-            return 3
-        else:
-            return 1
-
-    def _find_amounts_near_totals(self, lines: List[str]) -> List[Tuple[float, int, int, str]]:
-        """Find amounts near total keywords within 2 lines."""
-        nearby_amounts = []
-        
-        for i, line in enumerate(lines):
-            if any(keyword in line.lower() for keyword in ['grand total', 'total:', 'amount due']):
-                for offset in [-2, -1, 1, 2]:
-                    if 0 <= i + offset < len(lines):
-                        nearby_line = lines[i + offset]
-                        amounts = self.extract_amount_from_line(nearby_line)
-                        for amount in amounts:
-                            nearby_amounts.append((amount, 4, i + offset, nearby_line))
-                            self._debug_print(f"Found near total: ₹{amount} in line {i + offset}")
-        
-        return nearby_amounts
-
-    def _select_best_amount(self, amounts_found: List[Tuple[float, int, int, str]]) -> float:
-        """Select the best amount from candidates."""
+        # Debug: Show all amounts found
         if self.debug and len(amounts_found) > 1:
             print("\nDEBUG: All amounts found (sorted by priority):")
             sorted_amounts = sorted(amounts_found, key=lambda x: (x[1], x[0]), reverse=True)
-            for amt, pri, line_num, line_text in sorted_amounts[:5]:
+            for amt, pri, line_num, line_text in sorted_amounts[:5]:  # Show top 5
                 print(f"  Priority {pri}: ₹{amt} at line {line_num}: {line_text[:60]}...")
         
-        # Sort by priority then amount
+        # Sort by priority (highest first), then by amount (largest first)
         amounts_found.sort(key=lambda x: (x[1], x[0]), reverse=True)
-        selected = amounts_found[0]
         
+        selected = amounts_found[0]
         self._debug_print(f"Selected: ₹{selected[0]} from line {selected[2]}: {selected[3][:60]}...")
         
-        # Validation for large amounts
+        # Validation: For amounts over 1000, double-check if it makes sense
         if selected[0] > 1000:
+            # Check if there's a more reasonable amount with high priority
             high_priority_amounts = [a for a in amounts_found if a[1] >= 4 and a[0] < 1000]
-            if high_priority_amounts and high_priority_amounts[0][1] == 5:  # Grand Total
-                selected = high_priority_amounts[0]
-                self._debug_print(f"Switched to: ₹{selected[0]} from line {selected[2]}")
+            if high_priority_amounts:
+                # Prefer the smaller amount if it has high priority (likely the real total)
+                alternative = high_priority_amounts[0]
+                self._debug_print(f"Large amount detected (₹{selected[0]}), considering alternative: ₹{alternative[0]}")
+                # Only switch if the alternative has priority 5 (Grand Total line)
+                if alternative[1] == 5:
+                    selected = alternative
+                    self._debug_print(f"Switched to: ₹{selected[0]} from line {selected[2]}")
         
         return selected[0]
-
     def categorize_receipt(self, text: str, amount: Optional[float] = None) -> Tuple[str, float]:
-        """Categorize receipt based on text content."""
+        """
+        Categorize receipt based on text content.
+        
+        Returns:
+            Tuple of (category_name, confidence_score)
+        """
         if not text:
             return "Uncategorized", 0.0
         
         text_lower = text.lower()
-        scores = {}
         
+        # Score each category
+        scores = {}
         for category, info in self.categories.items():
             score = sum(1 for keyword in info['keywords'] if keyword.lower() in text_lower)
             scores[category] = score
         
+        # Find best match
         if scores and max(scores.values()) > 0:
             best_category = max(scores, key=scores.get)
             confidence = scores[best_category] / len(self.categories[best_category]['keywords'])
             return best_category, confidence
         
         return "Uncategorized", 0.0
-
+    
     def process_receipt(self, image_path: str) -> Dict[str, Any]:
         """Process a single receipt and extract information."""
         filename = os.path.basename(image_path)
@@ -451,16 +474,26 @@ class ReceiptProcessor:
             print('='*60)
         
         try:
+            # Extract text
             text = self.extract_text(image_path)
             
             if self.debug and text:
-                print(f"\nExtracted text preview:\n{'-'*40}")
-                print(text[:500] + ("..." if len(text) > 500 else ""))
+                print("\nExtracted text preview:")
+                print('-'*40)
+                print(text[:500] + "..." if len(text) > 500 else text)
                 print('-'*40)
             
             if not text:
-                return self._create_error_result(filename, 'No text extracted')
+                return {
+                    'file': filename,
+                    'amount': None,
+                    'category': 'Error',
+                    'confidence': 0.0,
+                    'error': 'No text extracted',
+                    'extracted_text': ''
+                }
             
+            # Extract amount and categorize
             amount = self.extract_amount(text)
             category, confidence = self.categorize_receipt(text, amount)
             
@@ -470,31 +503,29 @@ class ReceiptProcessor:
                 'category': category,
                 'confidence': confidence,
                 'error': None,
-                'extracted_text': text[:200] + ('...' if len(text) > 200 else text)
+                'extracted_text': text[:200] + '...' if len(text) > 200 else text
             }
             
         except Exception as e:
+            error_msg = str(e)
             if self.debug:
-                print(f"\nError: {e}")
+                print(f"\nError: {error_msg}")
                 traceback.print_exc()
-            return self._create_error_result(filename, str(e))
-
-    def _create_error_result(self, filename: str, error: str) -> Dict[str, Any]:
-        """Create standardized error result."""
-        return {
-            'file': filename,
-            'amount': None,
-            'category': 'Error',
-            'confidence': 0.0,
-            'error': error,
-            'extracted_text': ''
-        }
-
+            
+            return {
+                'file': filename,
+                'amount': None,
+                'category': 'Error',
+                'confidence': 0.0,
+                'error': error_msg,
+                'extracted_text': ''
+            }
+    
     def process_batch(self, input_folder: str, output_csv: str):
         """Process all receipt images in a folder."""
         input_path = Path(input_folder)
         
-        # Find supported files
+        # Find all supported files
         image_files = []
         for ext in self.SUPPORTED_FORMATS:
             image_files.extend(input_path.glob(f'*{ext}'))
@@ -520,19 +551,18 @@ class ReceiptProcessor:
             amount_str = f"₹{result['amount']:,.2f}" if result['amount'] else "Not found"
             print(f"{status} Amount: {amount_str} | Category: {result['category']} ({result['confidence']:.0%})")
         
+        # Save results
         self.save_to_csv(results, output_csv)
-        self._print_summary(results, output_csv)
-
-    def _print_summary(self, results: List[Dict], output_csv: str):
-        """Print processing summary."""
-        successful = sum(1 for r in results if r['error'] is None)
-        print(f"\n{'='*50}")
+        
+        # Print summary
+        print("\n" + "="*50)
         print("PROCESSING COMPLETE")
-        print(f"{'='*50}")
+        print("="*50)
+        successful = len([r for r in results if r['error'] is None])
         print(f"✓ Successful: {successful}/{len(results)}")
         print(f"✗ Failed: {len(results) - successful}/{len(results)}")
         print(f"📁 Results saved to: {output_csv}")
-
+    
     def save_to_csv(self, results: List[Dict], output_file: str):
         """Save processing results to CSV file."""
         if not results:
@@ -546,10 +576,12 @@ class ReceiptProcessor:
             writer.writeheader()
             
             for result in results:
+                # Format for CSV
                 row = result.copy()
                 if row['amount'] is not None:
                     row['amount'] = f"{row['amount']:.2f}"
                 row['confidence'] = f"{row['confidence']:.1%}"
+                
                 writer.writerow(row)
 
 
@@ -567,11 +599,16 @@ Examples:
         """
     )
     
-    parser.add_argument('input_folder', help='Folder containing receipt images')
-    parser.add_argument('-o', '--output', default='receipts_processed.csv',
+    parser.add_argument('input_folder', 
+                       help='Folder containing receipt images')
+    parser.add_argument('-o', '--output', 
+                       default='receipts_processed.csv',
                        help='Output CSV file (default: receipts_processed.csv)')
-    parser.add_argument('--debug', action='store_true', help='Enable debug output')
-    parser.add_argument('--vision-api', action='store_true', 
+    parser.add_argument('--debug', 
+                       action='store_true',
+                       help='Enable debug output')
+    parser.add_argument('--vision-api', 
+                       action='store_true',
                        help='Use Google Vision API (requires setup)')
     
     args = parser.parse_args()
@@ -588,9 +625,13 @@ Examples:
     
     # Initialize and run processor
     try:
-        processor = ReceiptProcessor(use_vision_api=args.vision_api, debug=args.debug)
+        processor = ReceiptProcessor(
+            use_vision_api=args.vision_api,
+            debug=args.debug
+        )
         processor.process_batch(args.input_folder, args.output)
         return 0
+        
     except KeyboardInterrupt:
         print("\n\nProcess interrupted by user")
         return 1
